add registry helpers to expand and identify sources
